Move cli apps out of internal packages
so that they can be used by other packages or commands, for example the docker entrypoint.
