+++ cmd/bert/app/encode.go
